Entity 소분류
