+++ src/main/java/com/projects/jslarticle/entity/admin/Admin.java
@@ -1,5 +1,7 @@
-package com.projects.jslarticle.entity;
+package com.projects.jslarticle.entity.admin;
 
+import com.projects.jslarticle.entity.board.Board;
+import com.projects.jslarticle.entity.user.User;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -9,29 +11,21 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToOne;
 import jakarta.persistence.Table;
-import jakarta.persistence.UniqueConstraint;
 import java.time.LocalDateTime;
 import lombok.Getter;
 
 /**
  * @author 탁영복
  * @version 1.0.0
- * @description ArticleCategory Entity 입니다. 추가 Entity 제작후 수정필요 합니다.
- * @since 2025-08-27
+ * @description Admin Entity 입니다. 추가 Entity 제작후 수정필요 합니다.
+ * @since 2025-08-29
  */
-@Table(
-        name = "article_category",
-        uniqueConstraints = {
-                @UniqueConstraint(
-                        name = "uk_article_category_board_id_category_name",
-                        columnNames = {"board_id", "category_name"})
-        }
-)
+@Table(name = "admin")
 @Entity
 @Getter
-public class ArticleCategory {
-
+public class Admin {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
@@ -39,27 +33,38 @@ public class ArticleCategory {
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(
             name = "board_id",
-            nullable = false,
-            foreignKey = @ForeignKey(name = "fk_article_category_board_id")
+            foreignKey = @ForeignKey(name = "fk_admin_board_id")
     )
     private Board board;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(
+            name = "user_id",
+            nullable = false,
+            foreignKey = @ForeignKey(name = "fk_admin_user_id")
+    )
+    private User user;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(
             name = "role_id",
             nullable = false,
-            foreignKey = @ForeignKey(name = "fk_article_category_role_id")
+            foreignKey = @ForeignKey(name = "fk_admin_role_id")
     )
     private Role role;
 
-    @Column(nullable = false, length = 20)
-    private String categoryName;
-
-    @Column(nullable = false)
-    private Integer sortOrder;
+    @OneToOne(fetch = FetchType.LAZY)
+    @JoinColumn(
+            name = "admin_config_id",
+            nullable = false,
+            foreignKey = @ForeignKey(name = "fk_admin_admin_config_id")
+    )
+    private AdminConfig adminConfig;
 
     @Column(updatable = false, nullable = false)
     private LocalDateTime createdAt;
 
     private LocalDateTime updatedAt;
+
+    private LocalDateTime deletedAt;
 }
