+++ src/main/java/com/projects/jslarticle/entity/article/Article.java
@@ -1,5 +1,7 @@
-package com.projects.jslarticle.entity;
+package com.projects.jslarticle.entity.article;
 
+import com.projects.jslarticle.entity.content.Content;
+import com.projects.jslarticle.entity.admin.Role;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
