+++ src/main/java/com/projects/jslarticle/entity/article/ArticleCategory.java
@@ -1,5 +1,7 @@
-package com.projects.jslarticle.entity;
+package com.projects.jslarticle.entity.article;
 
+import com.projects.jslarticle.entity.board.Board;
+import com.projects.jslarticle.entity.admin.Role;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -9,21 +11,29 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToOne;
 import jakarta.persistence.Table;
+import jakarta.persistence.UniqueConstraint;
 import java.time.LocalDateTime;
 import lombok.Getter;
 
 /**
  * @author 탁영복
  * @version 1.0.0
- * @description Admin Entity 입니다. 추가 Entity 제작후 수정필요 합니다.
- * @since 2025-08-29
+ * @description ArticleCategory Entity 입니다. 추가 Entity 제작후 수정필요 합니다.
+ * @since 2025-08-27
  */
-@Table(name = "admin")
+@Table(
+        name = "article_category",
+        uniqueConstraints = {
+                @UniqueConstraint(
+                        name = "uk_article_category_board_id_category_name",
+                        columnNames = {"board_id", "category_name"})
+        }
+)
 @Entity
 @Getter
-public class Admin {
+public class ArticleCategory {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
@@ -31,38 +41,27 @@ public class Admin {
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(
             name = "board_id",
-            foreignKey = @ForeignKey(name = "fk_admin_board_id")
+            nullable = false,
+            foreignKey = @ForeignKey(name = "fk_article_category_board_id")
     )
     private Board board;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(
-            name = "user_id",
-            nullable = false,
-            foreignKey = @ForeignKey(name = "fk_admin_user_id")
-    )
-    private User user;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(
             name = "role_id",
             nullable = false,
-            foreignKey = @ForeignKey(name = "fk_admin_role_id")
+            foreignKey = @ForeignKey(name = "fk_article_category_role_id")
     )
     private Role role;
 
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(
-            name = "admin_config_id",
-            nullable = false,
-            foreignKey = @ForeignKey(name = "fk_admin_admin_config_id")
-    )
-    private AdminConfig adminConfig;
+    @Column(nullable = false, length = 20)
+    private String categoryName;
+
+    @Column(nullable = false)
+    private Integer sortOrder;
 
     @Column(updatable = false, nullable = false)
     private LocalDateTime createdAt;
 
     private LocalDateTime updatedAt;
-
-    private LocalDateTime deletedAt;
 }
