+++ src/main/java/com/projects/jslarticle/entity/article/Bookmark.java
@@ -1,5 +1,6 @@
-package com.projects.jslarticle.entity;
+package com.projects.jslarticle.entity.article;
 
+import com.projects.jslarticle.entity.user.User;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -8,63 +9,48 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
-import jakarta.persistence.Lob;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
+import jakarta.persistence.UniqueConstraint;
 import java.time.LocalDateTime;
 import lombok.Getter;
 
 /**
  * @author 탁영복
  * @version 1.0.0
- * @description Report Entity 입니다. 추가 Entity 제작후 수정필요 합니다.
+ * @description Bookmark Entity 입니다. 추가 Entity 제작후 수정필요 합니다.
  * @since 2025-08-29
  */
-@Table(name = "report")
+@Table(
+        name = "bookmark",
+        uniqueConstraints = {
+                @UniqueConstraint(name = "uk_bookmark_user_id_article_id", columnNames = {"user_id", "artilce_id"})
+        }
+)
 @Entity
 @Getter
-public class Report {
+public class Bookmark {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(
-            name = "content_id",
-            nullable = false,
-            foreignKey = @ForeignKey(name = "fk_report_content_id")
-    )
-    private Content content;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(
             name = "user_id",
             nullable = false,
-            foreignKey = @ForeignKey(name = "fk_report_user_id")
+            foreignKey = @ForeignKey(name = "fk_bookmark_user_id")
     )
     private User user;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(
-            name = "admin_id",
-            foreignKey = @ForeignKey(name = "fk_report_admin_id")
+            name = "article_id",
+            nullable = false,
+            foreignKey = @ForeignKey(name = "fk_bookmark_article_id")
     )
-    private Admin admin;
-
-    @Lob
-    @Column(nullable = false)
-    private String description;
-
-    @Column(length = 255)
-    private String imageUrl;
-
-    @Column(length = 255)
-    private String result;
-
-    private Boolean isProcessed;
+    private Article article;
 
     @Column(updatable = false, nullable = false)
     private LocalDateTime createdAt;
-
-    private LocalDateTime updatedAt;
 }
