+++ src/main/java/com/projects/jslarticle/entity/board/BoardConfig.java
@@ -1,4 +1,4 @@
-package com.projects.jslarticle.entity;
+package com.projects.jslarticle.entity.board;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
