+++ src/main/java/com/projects/jslarticle/entity/content/Report.java
@@ -1,15 +1,16 @@
-package com.projects.jslarticle.entity;
+package com.projects.jslarticle.entity.content;
 
+import com.projects.jslarticle.entity.admin.Admin;
+import com.projects.jslarticle.entity.user.User;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.ForeignKey;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
+import jakarta.persistence.Lob;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import java.time.LocalDateTime;
@@ -18,14 +19,13 @@ import lombok.Getter;
 /**
  * @author 탁영복
  * @version 1.0.0
- * @description Point Entity 입니다. 추가 Entity 제작후 수정필요 합니다.
- * @since 2025-09-02
+ * @description Report Entity 입니다. 추가 Entity 제작후 수정필요 합니다.
+ * @since 2025-08-29
  */
-@Table(name = "point")
+@Table(name = "report")
 @Entity
 @Getter
-public class Point {
-
+public class Report {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
@@ -34,7 +34,7 @@ public class Point {
     @JoinColumn(
             name = "content_id",
             nullable = false,
-            foreignKey = @ForeignKey(name = "fk_point_content_id")
+            foreignKey = @ForeignKey(name = "fk_report_content_id")
     )
     private Content content;
 
@@ -42,20 +42,31 @@ public class Point {
     @JoinColumn(
             name = "user_id",
             nullable = false,
-            foreignKey = @ForeignKey(name = "fk_point_user_id")
+            foreignKey = @ForeignKey(name = "fk_report_user_id")
     )
     private User user;
 
-    @Column(nullable = false, length = 20)
-    @Enumerated(EnumType.STRING)
-    private ContentType contentType;
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(
+            name = "admin_id",
+            foreignKey = @ForeignKey(name = "fk_report_admin_id")
+    )
+    private Admin admin;
 
+    @Lob
     @Column(nullable = false)
-    private Integer changeAmount;
-
-    @Column(nullable = false, length = 255)
     private String description;
+
+    @Column(length = 255)
+    private String imageUrl;
+
+    @Column(length = 255)
+    private String result;
+
+    private Boolean isProcessed;
 
     @Column(updatable = false, nullable = false)
     private LocalDateTime createdAt;
+
+    private LocalDateTime updatedAt;
 }
