+++ src/main/java/com/projects/jslarticle/entity/emoji/EmojiUser.java
@@ -1,5 +1,7 @@
-package com.projects.jslarticle.entity;
+package com.projects.jslarticle.entity.emoji;
 
+import com.projects.jslarticle.entity.user.User;
+import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.ForeignKey;
@@ -9,18 +11,19 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
+import java.time.LocalDateTime;
 import lombok.Getter;
 
 /**
  * @author 탁영복
  * @version 1.0.0
- * @description Ban Entity 입니다. 추가 Entity 제작후 수정필요 합니다.
- * @since 2025-09-02
+ * @description EmojiUser Entity 입니다. 추가 Entity 제작후 수정필요 합니다.
+ * @since 2025-09-03
  */
-@Table(name = "ban")
+@Table(name = "emoji_user")
 @Entity
 @Getter
-public class Ban {
+public class EmojiUser {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -28,29 +31,23 @@ public class Ban {
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(
-            name = "admin_id",
+            name = "emoji_id",
             nullable = false,
-            foreignKey = @ForeignKey(name = "fk_ban_admin_id")
+            foreignKey = @ForeignKey(name = "fk_emoji_user_emoji_id")
     )
-    private Admin admin;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(
-            name = "board_id",
-            nullable = false,
-            foreignKey = @ForeignKey(name = "fk_ban_board_id")
-    )
-    private Board board;
+    private Emoji emoji;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(
             name = "user_id",
             nullable = false,
-            foreignKey = @ForeignKey(name = "fk_ban_user_id")
+            foreignKey = @ForeignKey(name = "fk_emoji_user_user_id")
     )
     private User user;
 
-    private String content;
+    @Column(updatable = false, nullable = false)
+    private LocalDateTime createdAt;
 
-    private Boolean isBlocked;
+    @Column(nullable = false)
+    private Integer sortOrder;
 }
