+++ src/main/java/com/projects/jslarticle/entity/user/Point.java
@@ -1,7 +1,11 @@
-package com.projects.jslarticle.entity;
+package com.projects.jslarticle.entity.user;
 
+import com.projects.jslarticle.entity.content.Content;
+import com.projects.jslarticle.entity.content.ContentType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.ForeignKey;
 import jakarta.persistence.GeneratedValue;
@@ -10,25 +14,19 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
-import jakarta.persistence.UniqueConstraint;
 import java.time.LocalDateTime;
 import lombok.Getter;
 
 /**
  * @author 탁영복
  * @version 1.0.0
- * @description Subscribe Entity 입니다. 추가 Entity 제작후 수정필요 합니다.
+ * @description Point Entity 입니다. 추가 Entity 제작후 수정필요 합니다.
  * @since 2025-09-02
  */
-@Table(
-        name = "role",
-        uniqueConstraints = {
-                @UniqueConstraint(name = "uk_subscribe_user_id_board_id", columnNames = {"user_id", "board_id"})
-        }
-)
+@Table(name = "point")
 @Entity
 @Getter
-public class Subscribe {
+public class Point {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -36,25 +34,29 @@ public class Subscribe {
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(
-            name = "user_id",
+            name = "content_id",
             nullable = false,
-            foreignKey = @ForeignKey(name = "fk_subscribe_user_id")
+            foreignKey = @ForeignKey(name = "fk_point_content_id")
     )
-    private User user;
+    private Content content;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(
-            name = "board_id",
+            name = "user_id",
             nullable = false,
-            foreignKey = @ForeignKey(name = "fk_subscribe_board_id")
+            foreignKey = @ForeignKey(name = "fk_point_user_id")
     )
-    private Board board;
+    private User user;
+
+    @Column(nullable = false, length = 20)
+    @Enumerated(EnumType.STRING)
+    private ContentType contentType;
 
     @Column(nullable = false)
-    private Integer sortOrder;
+    private Integer changeAmount;
 
-    @Column(nullable = false)
-    private Boolean notification;
+    @Column(nullable = false, length = 255)
+    private String description;
 
     @Column(updatable = false, nullable = false)
     private LocalDateTime createdAt;
